+++ controllers/ownPets/index.js
@@ -1,0 +1,9 @@
+const {addOwnPet} = require('./addOwnPet');
+const {deleteOwnPet} = require('./deleteOwnPet');
+const{listOwnPets} = require('./listOwnPetsAndUserInfo');
+
+module.exports = {
+    addOwnPet,
+    deleteOwnPet,
+    listOwnPets
+};
